Remove old stuff - Remove core/bot_utility.py completely - Remove debug/
