--- debug/db/summoners/summoner_db.dir
@@ -1,8 +1,0 @@
-'kryptux alias marc', (12800, 12203)
-'stormer', (25088, 16382)
-'susannova', (91136, 12398)
-'rantanplan', (103936, 8796)
-'Cyto', (113152, 13172)
-'Joey', (140800, 10290)
-'Truxxes', (151552, 14231)
-'maxteria', (165888, 16303)
