--- debug/riot-json/league.json
@@ -1,32 +1,0 @@
-[
-    {
-        "queueType": "RANKED_FLEX_SR",
-        "summonerName": "Thyanin",
-        "hotStreak": false,
-        "wins": 211,
-        "veteran": true,
-        "losses": 177,
-        "rank": "I",
-        "tier": "PLATINUM",
-        "inactive": false,
-        "freshBlood": false,
-        "leagueId": "2ac756a0-d259-11e9-978d-c81f66dd2a8f",
-        "summonerId": "GSei-TXNfq7IuwtmgiUDVOL-_Df3EFLPX83mL996JGODdK4",
-        "leaguePoints": 17
-    },
-    {
-        "queueType": "RANKED_SOLO_5x5",
-        "summonerName": "Thyanin",
-        "hotStreak": false,
-        "wins": 239,
-        "veteran": true,
-        "losses": 214,
-        "rank": "IV",
-        "tier": "DIAMOND",
-        "inactive": false,
-        "freshBlood": false,
-        "leagueId": "4ec60120-643e-11e9-96b3-c81f66dd0e0d",
-        "summonerId": "GSei-TXNfq7IuwtmgiUDVOL-_Df3EFLPX83mL996JGODdK4",
-        "leaguePoints": 19
-    }
-]
