--- debug/riot-json/summoner.json
@@ -1,9 +1,0 @@
-{
-    "profileIconId": 579,
-    "name": "Thyanin",
-    "puuid": "zNrwzL2GbDHjzRjMWBOiutXyFXBPKgiLHh_NHw6s7v7ua4rAsLWqa1dBI1tSvDMiLFpUIwMxQ7kyog",
-    "summonerLevel": 119,
-    "accountId": "oRrMhrf9gwtUwCjrAeWPL9enwLckSBDYUXzFZa3qcB2a0w",
-    "id": "GSei-TXNfq7IuwtmgiUDVOL-_Df3EFLPX83mL996JGODdK4",
-    "revisionDate": 1575586773000
-}
